20201013 sunny 询价最低价

+++ src/main/java/com/xzx/xzxms/vm/FinallyCheckCompareVM.java
@@ -2,14 +2,14 @@ package com.xzx.xzxms.vm;
 
 public class FinallyCheckCompareVM extends InquiryCompareVM {
 
-    //最低价
-    private Double minPrice;
+    //最低价(1代表最低价)
+    private Integer minPrice;
 
-    public Double getMinPrice() {
+    public Integer getMinPrice() {
         return minPrice;
     }
 
-    public void setMinPrice(Double minPrice) {
+    public void setMinPrice(Integer minPrice) {
         this.minPrice = minPrice;
     }
 }
